refactor(ui): extract GraphFlow behaviour hooks + dedup trace/view helpers
Lift four self-contained concerns out of the 1014-line graph-flow.tsx
component into focused hooks, and collapse two duplicated blocks. No
rendered-output change: the extracted hooks keep identical state, effects,
and (stable) dependency wiring.

- use-graph-context-menu.ts: ctxMenu state + dismiss-on-click/Escape effect.
- use-graph-search.ts: Fuse index + debounced query + result navigation.
- use-community-filter.ts: active communities + dimming + legend toggles.
- use-graph-keyboard-shortcuts.ts: the global keydown effect.
- graph-flow-helpers.ts: traceToEdgeKeys() was duplicated in the
  execution-flow effect and handlePathFound. The enter-full-view-from-module
  block (also duplicated in both) is now a single useCallback.

graph-flow.tsx 1014 -> 858 lines. tsc --noEmit clean; vitest 260 passed.
